Implement Redis Manager for client status managementThis commit introduces a new Redis manager module that handles client status management and related operations. Key features include:- Health checks for Redis connectivity with logging.- Automatic reconnection attempts to Redis with exponential backoff.- Synchronization of in-memory cache to Redis when available.- Cleanup of disconnected clients from both Redis and in-memory cache.- Functions to update and retrieve client statuses.Additionally, the server module has been refactored to utilize the new Redis manager, removing redundant code and improving maintainability.
